Makefile,README: host runs all make commands
Use `lxc exec` to invoke commands on the testing VM without need to open
a terminal on the VM. As such, there is no need to copy the Makefile to
the test VM, and invoking any of the `make` directives has no risk of
damaging the host system.

Furthermore, consolidate all setup into a single `make prepare-vm`
command which creates or starts the test VM, installs all dependencies,
enables prompting, and updates the test snap and the prompting client in
the VM. Thus, there is no need for a separate `bootstrap-vm.sh` script.

To ensure that the VM is running without updating the test snap and
prompting client, `make create-vm` can still be used, and it is now (for
the most part) idempotent.

Additionally, `make integration-tests` now depends on `prepare-vm`, so
can be run from a freshly cloned project.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>