UDENG-2998 mvp command line client to support spread tests
The feature set of this client is pretty minimal to start with but
should allow for some simple interactions with apparmor prompts within
scripted tests such as those found in snapd's spread tests.

There is not currently an ability to script replying to multiple prompts
in succession but this should be sufficient to get us started.