test(aa_prompting_client): add widget tests for home prompt
Instead of closing the prompt via `exit(0)` we're using
`YaruWindow.of(context).close()` now, which is easier to test.
To pull in yaru_test, yaru needs to track an acutal version instead of a
git branch (which is generally a good idea now that we're working
towards a more stable release).