fix: connect client with handler service before enabling prompting
There must be a snap with the snap-interfaces-requests-control interface
connected and a "handler-service" attribute defined to be one of its
daemon apps before the apparmor prompting experimental feature can be
enabled. Also, if snapd fails to start the app defined as the handler
service, it will fail to enable the flag.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>