Fixing issue with libhandy on non-GNOME x11
Modifying the way we set up the GTK window using libhandy to match how
things are done in the app-center. The commit that set things up in the
app-center can be found here:
  https://github.com/ubuntu/app-center/commit/43e7ae0a6ce82706b77587899ad06c36a2fac781

Note that this is _not_ how the docs for the handy_window package tell
you to set things up (this is what we followed originally). Prior to
this change we tested the current code under multiple desktop
environments on Ubuntu 24.04 with the following results:

  - GNOME Wayland: works
  - GNOME X11: works
  - Penrose X11: uninitialised window, closing leave the process running
    and eating CPU.
  - i3: uninitialised window, closing leave the process running
    and eating CPU.

With the change all of the above desktop environments work as expected.