refactor: remove tpm state enum
There was some confusion about if non-TPM FDE states should be supported by the
disk encryption page. The answer for now is no, so I am reverting to using
AuthMode as the provider enum.

Additionally this pr adds more of the error handling we will need for messaging
the user about issues with state we can't support on thier system.